Fix malformed configuration table in docs
A change to correct some configuration names cause the configuration
table to be malformed, and thus not appearing in the output of the
documentation. Sphinx raises the following during the docs build:

    WARNING: Malformed table.
    Text in column margin in table line 29.

This change adds some spacing to properly align the table columns so
there's no overlow. With this change,
https://flask-limiter.readthedocs.io/en/latest/#configuration would be
properly rendered.