Skip initialization sequence when disabled
If the rate limiter is disabled using config, the init_app sequence should be
skipped entirely.

Additionally if the initialization sequence has not completed, the extension
can still be triggered even though the before_request callback isn't
registered when rate limit decorators wrap routes or blueprints. For
this scenario the guard that checks self.initialized is introduced.

Addresses #258