Ensure non view function decorations cooperate with defaults
To ensure that rate limit decorations on non view functions
that are configured to either override or merge with the defaults
behave consistently with their view function counterparts
hints have to be recorded at runtime since the decoration doesn't
necessarily happen in a static context.

In the below example, there is at least one limit that doesn't override the defaults
explicitly and therefore the default limit of 1/second should be
exercised. However we can only know the nature of the limits inside the
view function once it has been executed to completion once. The first
pass on this request will result in the before_request filter exercising
the default limits without knowing anything about the inner limits. This
is incorrect since the inner limits may have all decided to
override the defaults - but we'll have to live with that and correct
later. Once the view is entered the immediate deductions due to the
decorators on _l1 and _l3 will skip the defaults because they override default
however _l2 will want to include the default limits - however this will
be deduplicated with a per request "seen" list. Once the request has
completed a hint map associating the test endpoint with the limits
registered on _l1,_l2 & _l3 will be populated which will be used in
subsequent requests to avoid deducting the default limits in the
before_request phase and instead the inline deduction due to _l3
will be the one to include the default limit (if _l3 is called
conditionally, then the default limit would not end up being exercised
which would ... sigh ... technically be correct)

   limiter = Limiter(...., default_limits=["1/second"])

   @limiter.limit("3/day")
   def _l3():
      ...
   @app.route("/test")
   def test():
      @limiter.limit("1/day", override_defaults=True)
      def _l1():
         ...
      @limiter.limit("2/day", override_defaults=False)
      def _l2():
         ...
      return _l1() + _l2() + _l3()