Allow multiple limiters without key_prefix to exist
If multiple limiter instances are used to decorate a single route
each of the decorations/instances should be able to deduct their limits
completely independently as there never was a precidence of multiple
limiters being aware of each other.

This does have potential breaking changes as demonstrated by the need
to update a test case but the test case seemed to have been
demonstrating a gap rather than an ideal expected behavior.