Add gcp func to contrib.util
GCP handles each request throw different load balancers and reverse proxies so using remote_addr is unreliable to obtain the client IP. As per [their documentation](https://cloud.google.com/load-balancing/docs/https#x-forwarded-for_header), the client IP is carried on in the X-Forwarded-For header. However, this is an array where each network node of GCP appends it's IP so the first in the list is the client IP (unless it's a spoofed request).